+++ src-tauri/src/audio/mod.rs
@@ -10,6 +10,7 @@ mod decode;
 mod dev_io;
 pub mod device_commands;
 pub mod mix_commands;
+pub mod preload_commands;
 pub mod radio_commands;
 pub mod transport_commands;
 mod device_watcher;
